+++ src/Spectrograph.tsx
@@ -1,26 +1,16 @@
+import Tippy from '@tippyjs/react';
 import classNames from 'classnames';
 import React from 'react';
+import { hideAll } from 'tippy.js';
 
 import { ChartTooltip } from './ChartTooltip';
 import { ColorTableRow } from './ColorTableRow';
 import { ColorScale } from './stores/DatasetStore';
 import { TooltipTable } from './TooltipTable';
-import { Trigger } from './Trigger';
 import { formatNumber } from './util/numbers';
 import { ValueFormatter } from './util/units/valueFormats';
 
 import styles from './Spectrograph.less';
-
-const BUILT_IN_PLACEMENTS = {
-  bottomLeft: {
-    points: ['tl', 'bl'],
-    offset: [0, 0],
-    overflow: {
-      adjustX: 1,
-      adjustY: 1,
-    },
-  },
-};
 
 export interface SpectrographProps {
   className?: string;
@@ -45,7 +35,7 @@ export class Spectrograph extends React.Component<SpectrographProps> {
         })?.color || defaultColor,
     }));
 
-    const rows = [];
+    const rows: JSX.Element[] = [];
 
     // Iterate backwards for the tooltip rows so the largest 'to' value is on top
     for (let i = valuesWithColor.length - 1; i >= 0; i -= 1) {
@@ -55,23 +45,26 @@ export class Spectrograph extends React.Component<SpectrographProps> {
       rows.push(<ColorTableRow key={`${i}-${count}`} name={`< ${name}`} values={[count]} color={color} />);
     }
 
-    const popupElement = (
-      <ChartTooltip>
-        <TooltipTable>
-          <tbody>{rows}</tbody>
-        </TooltipTable>
-      </ChartTooltip>
-    );
-
     return (
-      <Trigger
-        action={['hover']}
-        popupPlacement="bottomLeft"
-        builtinPlacements={BUILT_IN_PLACEMENTS}
-        popupTransitionName="slide-up"
-        popup={popupElement}
-        popupClassName={styles.tooltip}
-        destroyPopupOnHide
+      <Tippy
+        interactive={true}
+        placement="right"
+        theme="white"
+        trigger="mouseenter focus click"
+        content={
+          <ChartTooltip>
+            <TooltipTable>
+              <tbody>{rows}</tbody>
+            </TooltipTable>
+          </ChartTooltip>
+        }
+        onShow={(instance) => {
+          // Since we added `click` to the `trigger`, we need to
+          // hide all open Tippy's when a new one is shown.
+          hideAll({
+            exclude: instance,
+          });
+        }}
       >
         <div className={classNames(styles.spectrograph, className)}>
           {valuesWithColor.map((bucket: any, bucketIdx: number) => (
@@ -84,7 +77,7 @@ export class Spectrograph extends React.Component<SpectrographProps> {
             ></div>
           ))}
         </div>
-      </Trigger>
+      </Tippy>
     );
   }
 }