+++ CHANGELOG.md
@@ -6,3 +6,10 @@ All notable changes to this project will be documented in this file. See [standa
 ### 1.0.0-alpha.15 (2022-03-16)
 
 First "working" 🤞️ release to NPM.
+
+
+### 1.0.0-alpha.16 (2022-04-12)
+
+Added support for formatting numbers using the settings associated with each field.
+
+Replaced `rc-tooltip` with Tippy JS.